+++ app/src/main/java/com/example/pablo/globalfood/Model/MyRecipe.kt
@@ -5,19 +5,19 @@ import android.os.Parcelable
 
 @Suppress("NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS")
 
-data class MyRecipe (var title:String, var country:String, var resDish:String, var esFav:Int): Parcelable {
+data class MyRecipe (var title:String, var country:String, var resDish:String, var esFav:Long): Parcelable {
     constructor(parcel: Parcel) : this(
             parcel.readString(),
             parcel.readString(),
             parcel.readString(),
-            parcel.readInt()
+            parcel.readLong()
     )
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         parcel.writeString(title)
         parcel.writeString(country)
         parcel.writeString(resDish)
-        parcel.writeInt(esFav)
+        parcel.writeLong(esFav)
     }
 
     override fun describeContents(): Int {
